requirements.txt
lxml
tqdm
absl-py
easydict
matplotlib
pillow
opencv-python
tensorflow
